+++ integrations/azure-devops/integration.py
@@ -1,4 +1,4 @@
-from pydantic import Field, BaseModel, validator
+from pydantic import Field, BaseModel
 from port_ocean.context.ocean import PortOceanContext
 from port_ocean.core.handlers.port_app_config.api import APIPortAppConfig
 from port_ocean.core.handlers.webhook.processor_manager import (